requirements.txt
Flask>=3.0.0
treys>=0.1.8
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
